gateway/requirements.txt
# Cliente gRPC
grpcio

# Peticiones HTTP
requests

# Cliente MQTT
paho-mqtt

# Cliente WebSocket
websockets